% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{tmb}
\alias{tmb}
\title{tmb}
\usage{
tmb(signature.name, signature, cutoff.freq = 10, cutoff.percentile = 0.2)
}
\arguments{
\item{signature.name}{Names of a signature (format: character)}

\item{signature}{Gene names of a signature (format: vector)}

\item{cutoff.freq}{Cutoff for frequency of cancer cell lines for each cancer type, default 10}

\item{cutoff.percentile}{Cutoff for percentile of cancer cell lines with highest/lowest dependency, default 0.2}
}
\value{
plot
}
\description{
'tmb' compares the tumor mutation burden (TMB) between cancer cell lines with highest and lowest dependencies of a gene set (signature).
}
\examples{
source(system.file("script", "load_libs.R", package = "deplink"))
signature.name = "9-1-1"
signature = c("RAD9A", "RAD1", "HUS1", "RAD17")
tmb(signature.name, signature)
}
\references{
1. X Chen, J McGuire, F Zhu, X Xu, Y Li, D Karagiannis, R Dalla-Favera, A Ciccia, J Amengual, C Lu (2020). 
Harnessing genetic dependency correlation network to reveal chromatin vulnerability in cancer.  
In preparation.
}
\author{
Xiao Chen
}
